+++ ErsatzTV.Core/Jellyfin/JellyfinMediaStreamType.cs
@@ -1,0 +1,8 @@
+namespace ErsatzTV.Core.Jellyfin;
+
+public static class JellyfinMediaStreamType
+{
+    public static readonly string Video = "Video";
+    public static readonly string Audio = "Audio";
+    public static readonly string Subtitle = "Subtitle";
+}